+++ techgrains/test/com/techgrains/util/tg_file_util_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:techgrains/com/techgrains/util/tg_file_util.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -7,10 +6,4 @@ void main() {
   setUp(() {});
 
   tearDown(() {});
-
-  test('alphanumeric', () async {
-    String content =
-        await TGFileUtil.readFileAsString("assets/files/company.txt");
-    expect(content, "Techgrains Technologies");
-  });
 }